Support .markdown file extension when filtering
`.md` & `.markdown` are the most common extensions, and are the two
supported by GitHub. The longer `.markdown` is actually the one endorsed
by John Gruber, Markdown's creator.

There are others, as well[1], but I think these two are fine for now
since they're by far the most used and are what GitHub supports as
well.