Accept directories in include_paths
Without this, the engine would likely only end up parsing Markdown files
in the root directory. (Which, admittedly, is most Markdown files, but
not all).

The path-parsing change in issue emission was necesary because `mdl`
internally does some unusual string construction of file names when
given relative paths to analyze. This is the easiest way to get around
it.