Suppress Kramdown Warnings
`mdl`'s default behavior is to print any warnings from Kramdown[1], and to
print them to stdout.

These might be helpful warnings to emit to stderr ourselves, but for the
moment at least we should probably just not get them, since they are a
slightly different format from issues, and  would result in invalid
output from the engine.

[1]: https://github.com/mivok/markdownlint/blob/346bfceb7ff85cbbe483c5dea78c7712015a0421/lib/mdl.rb#L76-L81