Generate a unique fingerprint per issue
Provide a generated fingerprint to Code Climate to prevent collisions
when the same style issue is reported across multiple lines per file.

Right now the default Code Climate fingerprint algorithm is a
concatenation of the check name and path hashed. Unfortunately this
results in a fair amount of collisions for markdownlint as often a
single issue is present multiple times in a given file.

The fingerprint implemented will use the check name, path, and the
contents of the subject line of the issue with whitespace removed.