Swap POSIX::Spawn.popen4 for POSIX::Spawn::Child
We generally use it in other codebases. The interface seems simpler to
me, as well, and it doesn't have this indefinite blocking problem.

The underlying tool doesn't do structured output, so we're stuck parsing
the issues from the output. That means we can't rely on the exit code of
the child process to determine whether the engine failed or not (it's
tied up representing whether there are issues or not).

So, if the underlying tool DOES fail, we'll print the error output, but
we won't bubble it up as an E10, because we can't be sure if there was a
problem or not (sometimes printing error output is done for debugging,
not because there was a fatal problem).

It might happen that we'll then have a ruby error while we try to parse
the issue output, which would be great, because then the user will see
all of the output.

I think the best thing to do is to go implement structured output for
the underlying tool if we're not comfortable with that.